+++ src/actions/update.js
@@ -24,6 +24,9 @@ const {
   FILES_TAGS_INDEX_KEY,
   FILES_USER_INDEX_PUBLIC_KEY,
   FILES_PLAYER_SETTINGS_FIELD,
+  FILES_VERSION_FIELD,
+  FILES_NFT_FIELD,
+  FILES_HAS_NFT,
 } = require('../constant');
 
 const { call } = Function.prototype;
@@ -101,6 +104,14 @@ async function updateMeta(lock, ctx, params) {
   const userPublicIndex = FILES_USER_INDEX_PUBLIC_KEY(owner);
   const isPublic = data[FILES_PUBLIC_FIELD];
 
+  // update version
+  const nftImage = meta.nft && meta.nft.image;
+  const prevImage = data.nft && data.nft.image;
+
+  if (nftImage && nftImage !== prevImage) {
+    pipeline.hincrby(key, FILES_VERSION_FIELD, 1);
+  }
+
   if (alias) {
     pipeline.hset(aliasPTRs, alias, uploadId);
     if (existingAlias) {
@@ -156,12 +167,17 @@ async function updateMeta(lock, ctx, params) {
 
   // make sure it's not an empty object
   if (hasProperties(meta)) {
+    // to index using this property with redis-search
+    if (meta[FILES_NFT_FIELD]) {
+      meta[FILES_HAS_NFT] = '1';
+    }
+
     pipeline.hmset(key, meta);
   }
 
   handlePipeline(await pipeline.exec());
 
-  if (directOnly !== undefined) {
+  if (directOnly !== undefined || (meta.nft && !data.nft)) {
     await bustCache(redis, data, true, true);
   }