+++ src/constant.js
@@ -48,6 +48,8 @@ module.exports = exports = Object.setPrototypeOf({
   FILES_CONTENT_LENGTH_FIELD: 'contentLength',
   FILES_UPLOADED_AT_FIELD: 'uploadedAt',
   FILES_UPLOAD_STARTED_AT_FIELD: 'startedAt',
+  FILES_VERSION_FIELD: 'version',
+  FILES_HAS_NFT: 'hn',
 
   // metatype of file
   FILES_TYPE_FIELD: 'type',
@@ -79,6 +81,7 @@ exports.FILES_ARPROPS_FIELD = 'ar3dviewProps';
 exports.FILES_CREATION_INFO_FIELD = 'creationInfo';
 exports.FILES_PLAYER_SETTINGS_FIELD = 'playerSettings';
 exports.FILES_PLAYER_FRAMES_CYCLE_FIELD = 'cycle';
+exports.FILES_NFT_FIELD = 'nft';
 
 exports.FIELDS_TO_STRINGIFY = [
   exports.FILES_TAGS_FIELD,
@@ -90,6 +93,7 @@ exports.FIELDS_TO_STRINGIFY = [
   exports.FILES_CREATION_INFO_FIELD,
   exports.FILES_PLAYER_SETTINGS_FIELD,
   exports.FILES_PLAYER_FRAMES_CYCLE_FIELD,
+  exports.FILES_NFT_FIELD,
 ];
 
 exports.CAPPASITY_3D_MODEL = 'model';